cpu: fix prevent
Fix below prevent defects:

CID 310276
CID 276946
CID 239749
CID 239748
CID 239745
CID 239741
CID 239740
CID 239739
CID 239738
CID 239736
CID 239732
CID 214893
CID 310340
CID 310302
CID 310259
CID 310232
CID 310170
CID 310161
CID 280587
CID 277691
CID 277612
CID 277542
CID 277463
CID 276681
CID 276665
CID 239733
CID 233850
CID 233847
CID 233841
CID 233827
CID 233825
CID 233814

+++ drivers/cpufreq/exynos-acme.c
@@ -345,7 +345,7 @@ static int exynos_cpufreq_driver_init(struct cpufreq_policy *policy)
 static unsigned int exynos_cpufreq_resolve(struct cpufreq_policy *policy,
 						unsigned int target_freq)
 {
-	unsigned int index;
+	int index;
 
 	index = cpufreq_frequency_table_target(policy, target_freq, CPUFREQ_RELATION_L);
 	if (index < 0) {
@@ -732,9 +732,9 @@ static int exynos_cpufreq_pm_qos_callback(struct notifier_block *nb,
 	/* If 'sugov_get_freq()' fail, we just update frequency with pm qos val */
 	if (next_freq) {
 		if (pm_qos_class == domain->pm_qos_min_class)
-			val = max_t(unsigned int, val, next_freq);
+			val = max_t(unsigned long, val, next_freq);
 		else
-			val = min_t(unsigned int, val, next_freq);
+			val = min_t(unsigned long, val, next_freq);
 	}
 
 	if (update_freq(domain, val))
